requirements.txt
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
EasyProcess==1.1
entrypoint2==1.1
filelock==3.12.0
idna==3.4
lxml==4.9.2
MouseInfo==0.1.3
mss==9.0.1
numpy==1.24.3
pandas==2.0.2
Pillow==9.5.0
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyobjc-core==9.1.1
pyobjc-framework-Cocoa==9.1.1
pyobjc-framework-Quartz==9.1.1
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
PySocks==1.7.1
python-dateutil==2.8.2
pytweening==1.0.7
pytz==2023.3
requests==2.31.0
rubicon-objc==0.4.6
six==1.16.0
snscrape @ git+https://github.com/JustAnotherArchivist/snscrape.git@b76f485208016f337ce36a3c6c0c6573502693a7
soupsieve==2.4.1
tzdata==2023.3
urllib3==2.0.2
